mindlink/requirements.txt
# Mind Link — Python Dependencies
# Install: pip install -r requirements.txt

# Core scientific
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
pyyaml>=6.0

# EEG / BCI
brainflow>=5.10
mne>=1.5
pylsl>=1.16

# Quantum
qiskit>=1.0
qiskit-aer>=0.13
qiskit-machine-learning>=0.7

# Transmission
pymavlink>=2.4
bleak>=0.21

# Safety / TTS
pyttsx3>=2.90

# Visualization
matplotlib>=3.7
ursina>=6.0

# Testing
pytest>=7.4
